Move jobs into their own module
